+++ libdiskonaut/src/model/files/mod.rs
@@ -1,6 +1,7 @@
 mod file_or_folder;
 mod file_tree;
 mod hard_links;
+mod hash;
 
 pub use file_or_folder::*;
 pub use file_tree::*;